shell: Add defaultDexConnector option to automatically trigger login on a predefined Dex connector.
This option can be added to the auth section of micro application `runtime-app-configuration`.
It intercept the authorization_endpoint of the oidc web finger and append it the `connector_id` query parameter.

The connector choice can be forced by adding the `displayLoginChoice` query parameter to the current URL.
This enable to maintain UI login possible when the default connector is unavailable.